requirements.txt
# ============================================
# JCVI Genome Analyzer - Requirements
# ============================================
# Install Python packages:
#   pip install -r requirements.txt
#
# Install system-level tools (Linux):
#   sudo apt install ncbi-blast+
#   sudo apt install docker.io docker-compose   (only if using full OQTOPUS backend)
# ============================================

# --- Core (REQUIRED) ---
PyQt5>=5.15
biopython>=1.79
requests>=2.28

# --- AI & Data Science (Recommended) ---
numpy>=1.21
pandas>=1.3
torch>=1.9
openpyxl>=3.0

# --- Optional Features ---
pyzmq>=22.0          # Live ZeroMQ data broadcasting
flask>=2.0           # Standalone quantum API server
psutil>=5.9          # Benchmarking & system monitoring

# ============================================
# System-Level Tools (install separately)
# ============================================
# 1. NCBI BLAST+ (REQUIRED for genome comparison)
#    Ubuntu/Debian:  sudo apt install ncbi-blast+
#    Conda:          conda install -c bioconda blast
#    Verify:         blastn -version
#
# 2. Docker & Docker Compose (OPTIONAL - only for full OQTOPUS quantum backend)
#    Ubuntu/Debian:  sudo apt install docker.io docker-compose
#    Verify:         docker --version && docker-compose --version
# ============================================
